requirements.txt
click~=7.0
datasets~=1.1.3
numpy~=1.17.0
scikit-learn~=0.19.1
scipy~=1.3.1
tensorboardX~=1.7
torchsummary
tqdm
transformers~=4.1.0
